fix: ensure `make test` can be run
Prior to this commit, `make test` would fail because one of the Makefile
targets was invoking `./bin/controller-gen` with the wrong arguments.

To fix the `controller-gen` invocation, the empty `config/crd/bases`
directory had to be added.

Finally, the invocation of `controller-gen` has been changed to reflect
the one generated by latest version of kube-builder.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>